+++ controls/admin.py
@@ -1,8 +1,7 @@
 import csv
 from django.contrib import admin
 from django.http import HttpResponse
-from .models import ImportRecord, Statement, Element, ElementControl, System, CommonControlProvider, CommonControl, ElementCommonControl, Poam, Deployment
-# from .models import InventoryItemAssessmentResults
+from .models import ImportRecord, Statement, Element, ElementControl, System, CommonControlProvider, CommonControl, ElementCommonControl, Poam, Deployment, SystemAssessmentResult
 from guardian.admin import GuardedModelAdmin
 from simple_history.admin import SimpleHistoryAdmin
 
@@ -77,14 +76,8 @@ class DeploymentAdmin(SimpleHistoryAdmin, ExportCsvMixin):
     def uuid(self, obj):
         return obj.deployment.uuid
 
-# class InventoryItemAssessmentResultsAdmin(admin.ModelAdmin):
-#     list_display = ('id', 'statement')
-
-#     # def uuid(self, obj):
-#     #     return obj.statement.uuid
-
-#     # def consumer_element(self, obj):
-#     #     return obj.statement.consumer_element
+class SystemAssessmentResultAdmin(admin.ModelAdmin):
+    list_display = ('id', 'system', 'deployment')
 
 
 admin.site.register(ImportRecord, ImportRecordAdmin)
@@ -97,5 +90,5 @@ admin.site.register(CommonControl, CommonControlAdmin)
 admin.site.register(ElementCommonControl, ElementCommonControlAdmin)
 admin.site.register(Poam, PoamAdmin)
 admin.site.register(Deployment, DeploymentAdmin)
-# admin.site.register(InventoryItemAssessmentResults, InventoryItemAssessmentResultsAdmin)
+admin.site.register(SystemAssessmentResult, SystemAssessmentResultAdmin)