+++ siteapp/urls.py
@@ -4,6 +4,7 @@ from django.conf import settings
 
 admin.autodiscover()
 
+import controls.views_api
 import siteapp.views as views
 import siteapp.views_landing as views_landing
 import siteapp.views_health as views_health
@@ -29,6 +30,7 @@ urlpatterns = [
 
     # Controls and Systems
     url(r"^systems/", include("controls.urls")),
+    url(r"^api/v1/systems/", include("controls.urls_api")),
     url(r"^controls/", include("controls.urls")),
 
     # app store
@@ -39,7 +41,7 @@ urlpatterns = [
     url(r'^library/(?P<source_slug>.*)/(?P<app_name>.*)$', views.apps_catalog_item),
 
     # projects
-    url(r"^projects$", views.project_list, name="projects"),
+    url(r"^projects$", views.ProjectList.as_view(), name="projects"),
     url(r"^projects/lifecycle$", views.project_list_lifecycle, name="projects_lifecycle"),
     url(r'^projects/(\d+)/__rename$', views.rename_project, name="rename_project"),
     url(r'^projects/(\d+)/__delete$', views.delete_project, name="delete_project"),